+++ src/types/auth/tokenValid.ts
@@ -1,0 +1,9 @@
+import { AxiosResponse } from 'axios';
+
+interface CheckTokenValidData {
+  statusCode: number;
+  message?: string;
+  error?: string;
+}
+
+export type CheckTokenValidResponse = AxiosResponse<CheckTokenValidData>;
